antigravity-core/pyproject.toml
[project]
name = "antigravity-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "fpdf2>=2.8.7",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "polars>=1.38.1",
    "psutil>=7.2.2",
    "pyjwt[crypto]>=2.11.0",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.41.0",
]
